3er push gestion de productos

+++ tpo/src/main/java/com/uade/tpo/service/DescuentosServiceImplement.java
@@ -1,11 +1,12 @@
 package com.uade.tpo.service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.uade.tpo.entity.Descuentos;
@@ -18,41 +19,45 @@ public class DescuentosServiceImplement implements DescuentosService {
     private DescuentosRepository descuentosRepository;
 
     @Override
-    public Page<Descuentos> getDescuentos(PageRequest pageable) {
+    public Page<Descuentos> getDescuentos(Pageable pageable) {
         return descuentosRepository.findAll(pageable);
-        throw new UnsupportedOperationException("Unimplemented method 'getDescuentos'");
     }
 
     @Override
-    public Optional<Descuentos> getDescuentosById(Integer idDescuentos) {
-        return descuentosRepository.findById(idDescuentos).orElse(null);
-        throw new UnsupportedOperationException("Unimplemented method 'getDescuentosById'");
+    public Optional<Descuentos> getDescuentosById(Long idDescuentos) {
+        return descuentosRepository.findById(idDescuentos);
     }
 
     @Override
-    public Descuentos crearDescuentos(Integer idDescuentos) {
-        List<Descuentos> descuentos = descuentosRepository.findById(idDescuentos);
-        if (descuentos.isEmpty())
-            return descuentosRepository.save(new Descuentos(idDescuentos));
-        throw new UnsupportedOperationException("Unimplemented method 'crearDescuentos'");
+    public Descuentos crearDescuentos(Descuentos descuentos) {
+        Optional<Descuentos> listaDescuentos = descuentosRepository.findById(descuentos.getIdDescuentos());
+        if (listaDescuentos.isEmpty()){
+            return descuentosRepository.save(descuentos);
+        }else{
+            throw new IllegalStateException("El descuento ya existe");
+        }
     }
 
     @Override
     public List<Descuentos> getAllDescuentos() {
         return descuentosRepository.findAll();
-        throw new UnsupportedOperationException("Unimplemented method 'getAllDescuentos'");
     }
 
     @Override
-    public Descuentos eliminarDescuentos(Integer idDescuentos) {
-        descuentosRepository.deleteById(idDescuentos);
-        throw new UnsupportedOperationException("Unimplemented method 'eliminarDescuentos'");
+    public Descuentos eliminarDescuentos(Long idDescuentos) {
+        Optional<Descuentos> descuentosOptional = descuentosRepository.findById(idDescuentos);
+        if(descuentosOptional.isPresent()){
+            Descuentos descuentoEliminado = descuentosOptional.get();
+            descuentosRepository.deleteById(idDescuentos);
+            return descuentoEliminado;
+        }else{
+            throw new NoSuchElementException("El descuento no existe");
+        } 
     }
-
+    
     @Override
-    public Descuentos modificarDescuentos(Integer idDescuentos) {
-        return descuentosRepository.save(idDescuentos);
-        throw new UnsupportedOperationException("Unimplemented method 'modificarDescuentos'");
+    public Descuentos actualizarDescuentos(Long idDescuentos, Descuentos descuentosActualizados) {
+        descuentosActualizados.setIdDescuentos(idDescuentos);
+        return descuentosRepository.save(descuentosActualizados);
     }
-
 }
